Move KafkaConfigStorage to the o.a.k.cc.storage package.
